mcp-server/requirements.txt
//requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
anthropic
mcp
boto3
httpx